requirements.txt
flask
google
google-genai
gunicorn
python-dotenv